added confirmation window before deletion

+++ frontend/src/app/progress-tracker/entries/entries.component.ts
@@ -5,6 +5,8 @@ import { MatTableDataSource } from '@angular/material/table';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { EntryService } from '../entry.service'
+import { MatDialog } from '@angular/material/dialog';
+import { AlertDialogComponent } from 'src/app/shared/alert-dialog/alert-dialog.component'
 
 export interface entry {
   id: number;
@@ -19,7 +21,6 @@ export interface entry {
 })
 export class EntriesComponent implements OnInit {
 
-  private message:string;;
   displayedColumns: string[] = ['date', 'weight', 'actions'];
   data: entry[] = [];
   dataSource = new MatTableDataSource<entry>(this.data);
@@ -29,11 +30,12 @@ export class EntriesComponent implements OnInit {
   
   constructor(
     private snackBar: MatSnackBar,
-    private _entryService: EntryService
+    private _entryService: EntryService,
+    private _dialog: MatDialog
   ) {}
 
   ngOnInit() {
-    this._entryService.myObservable.subscribe(
+    this._entryService.entriesObservable.subscribe(
       (entries: entry[]) => {this.dataSource.data = entries}
       )
     this.dataSource.paginator = this.paginator;
@@ -45,14 +47,23 @@ export class EntriesComponent implements OnInit {
   });
 
   deleteEntry(id: string): void{
-    this._entryService.deleteEntry(id).subscribe(
-      body => {
-        this._entryService.getEntries();
-      }, 
-      error => {
-        this.snackBar.open('Error occured while deleting entry.', 'Dismiss', {duration: 6000})
+    const dialogRef = this._dialog.open(AlertDialogComponent, {
+      width: '350px',
+      data: {title: 'Confirm deletion', message: 'Do you really want to delete this entry?', btn_cancel: 'Cancel', btn_confirm: 'Delete'}
+    });
+    dialogRef.afterClosed().subscribe(confirmed => {
+      if (confirmed){
+        this._entryService.deleteEntry(id).subscribe(
+          body => {
+            this._entryService.getEntries();
+          }, 
+          error => {
+            this.snackBar.open('Error occured while deleting entry.', 'Dismiss', {duration: 6000})
+            console.log(error)
+          }
+        )
       }
-    )
+    });
   }
 
 }
